snap: stop testing addImplicitHooks directly.
