+++ snap/implicit_test.go
@@ -20,11 +20,9 @@
 package snap_test
 
 import (
-	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/interfaces/builtin"
 	"github.com/snapcore/snapd/release"
 	"github.com/snapcore/snapd/snap"
-	"github.com/snapcore/snapd/snap/snaptest"
 
 	. "gopkg.in/check.v1"
 )
@@ -71,176 +69,5 @@ func (s *InfoSnapYamlTestSuite) TestImplicitSlotsAreRealInterfaces(c *C) {
 	}
 	for _, ifaceName := range snap.ImplicitClassicSlotsForTests {
 		c.Check(known[ifaceName], Equals, true)
-	}
-}
-
-func (s *InfoSnapYamlTestSuite) TestAddImplicitHooksNoHooks(c *C) {
-	dirs.SetRootDir(c.MkDir())
-	defer func() { dirs.SetRootDir("") }()
-
-	yaml := `name: foo
-version: 1.0`
-	info := snaptest.MockSnap(c, yaml, &snap.SideInfo{Revision: snap.R(42)})
-
-	// Now load implicit hooks
-	snap.AddImplicitHooks(info)
-
-	// Verify that no hooks were loaded for this snap
-	c.Check(info.Hooks, HasLen, 0)
-}
-
-func (s *InfoSnapYamlTestSuite) TestAddImplicitHooksFromContainerNoHooks(c *C) {
-	dirs.SetRootDir(c.MkDir())
-	defer func() { dirs.SetRootDir("") }()
-
-	yaml := `name: foo
-version: 1.0`
-	path := makeTestSnap(c, yaml)
-
-	container, err := snap.Open(path)
-	c.Assert(err, IsNil)
-
-	// Now load implicit hooks
-	info := &snap.Info{Hooks: make(map[string]*snap.HookInfo)}
-	snap.AddImplicitHooksFromContainer(info, container)
-
-	// Verify that no hooks were loaded for this snap
-	c.Check(info.Hooks, HasLen, 0)
-}
-
-func (s *InfoSnapYamlTestSuite) TestAddImplicitHooksSingle(c *C) {
-	dirs.SetRootDir(c.MkDir())
-	defer func() { dirs.SetRootDir("") }()
-
-	yaml := `name: foo
-version: 1.0`
-	info := snaptest.MockSnapWithHooks(c, yaml, &snap.SideInfo{Revision: snap.R(42)}, []string{"test-hook"})
-
-	// Verify that no hooks have been loaded for this snap
-	c.Check(info.Hooks, HasLen, 0)
-
-	// Now load implicit hooks
-	snap.AddImplicitHooks(info)
-
-	// Verify that the `test-hook` hook has now been loaded, and that it has no
-	// associated plugs.
-	c.Check(info.Hooks, HasLen, 1)
-	verifyImplicitHook(c, info, "test-hook")
-}
-
-func (s *InfoSnapYamlTestSuite) TestAddImplicitHooksFromContainerSingle(c *C) {
-	dirs.SetRootDir(c.MkDir())
-	defer func() { dirs.SetRootDir("") }()
-
-	yaml := `name: foo
-version: 1.0`
-	path := makeTestSnapWithHooks(c, yaml, []string{"test-hook"})
-
-	container, err := snap.Open(path)
-	c.Assert(err, IsNil)
-
-	// Now load implicit hooks
-	info := &snap.Info{Hooks: make(map[string]*snap.HookInfo)}
-	snap.AddImplicitHooksFromContainer(info, container)
-
-	// Verify that the `test-hook` hook has now been loaded, and that it has no
-	// associated plugs.
-	c.Check(info.Hooks, HasLen, 1)
-	verifyImplicitHook(c, info, "test-hook")
-}
-
-func (s *InfoSnapYamlTestSuite) TestAddImplicitHooksMultiple(c *C) {
-	dirs.SetRootDir(c.MkDir())
-	defer func() { dirs.SetRootDir("") }()
-
-	yaml := `name: foo
-version: 1.0`
-	info := snaptest.MockSnapWithHooks(c, yaml, &snap.SideInfo{Revision: snap.R(42)}, []string{"hook1", "hook2"})
-
-	// Verify that no hooks have been loaded for this snap
-	c.Check(info.Hooks, HasLen, 0)
-
-	// Now load implicit hooks
-	snap.AddImplicitHooks(info)
-
-	// Verify that both hooks have now been loaded, and that neither have any
-	// associated plugs.
-	c.Check(info.Hooks, HasLen, 2)
-	verifyImplicitHook(c, info, "hook1")
-	verifyImplicitHook(c, info, "hook2")
-}
-
-func (s *InfoSnapYamlTestSuite) TestAddImplicitHooksFromContainerMultiple(c *C) {
-	dirs.SetRootDir(c.MkDir())
-	defer func() { dirs.SetRootDir("") }()
-
-	yaml := `name: foo
-version: 1.0`
-	path := makeTestSnapWithHooks(c, yaml, []string{"hook1", "hook2"})
-
-	container, err := snap.Open(path)
-	c.Assert(err, IsNil)
-
-	// Now load implicit hooks
-	info := &snap.Info{Hooks: make(map[string]*snap.HookInfo)}
-	snap.AddImplicitHooksFromContainer(info, container)
-
-	// Verify that both hooks have now been loaded, and that neither have any
-	// associated plugs.
-	c.Check(info.Hooks, HasLen, 2)
-	verifyImplicitHook(c, info, "hook1")
-	verifyImplicitHook(c, info, "hook2")
-}
-
-func (s *InfoSnapYamlTestSuite) TestAddImplicitHooksWithExplicitHook(c *C) {
-	dirs.SetRootDir(c.MkDir())
-	defer func() { dirs.SetRootDir("") }()
-
-	yaml := `name: foo
-version: 1.0
-hooks:
-  explicit:
-    plugs: [test-plug]`
-	info := snaptest.MockSnapWithHooks(c, yaml, &snap.SideInfo{Revision: snap.R(42)}, []string{"explicit", "implicit"})
-
-	// Verify that `explicit` has already been loaded due to its being in the YAML.
-	c.Assert(info.Hooks, HasLen, 1)
-	verifyExplicitHook(c, info, "explicit", []string{"test-plug"})
-
-	// Now load implicit hooks
-	snap.AddImplicitHooks(info)
-
-	// Verify that the `implicit` hook has now been loaded, and that it has no
-	// associated plugs. Also verify that the `explicit` hook is still valid.
-	c.Check(info.Hooks, HasLen, 2)
-	verifyImplicitHook(c, info, "implicit")
-	verifyExplicitHook(c, info, "explicit", []string{"test-plug"})
-}
-
-func verifyImplicitHook(c *C, info *snap.Info, hookName string) {
-	hook := info.Hooks[hookName]
-	c.Assert(hook, NotNil, Commentf("Expected hooks to contain %q", hookName))
-	c.Check(hook.Name, Equals, hookName)
-	c.Check(hook.Plugs, IsNil)
-}
-
-func verifyExplicitHook(c *C, info *snap.Info, hookName string, plugNames []string) {
-	hook := info.Hooks[hookName]
-	c.Assert(hook, NotNil, Commentf("Expected hooks to contain %q", hookName))
-	c.Check(hook.Name, Equals, hookName)
-	c.Check(hook.Plugs, HasLen, len(plugNames))
-
-	for _, plugName := range plugNames {
-		// Verify that the HookInfo and PlugInfo point to each other
-		plug := hook.Plugs[plugName]
-		c.Assert(plug, NotNil, Commentf("Expected hook plugs to contain %q", plugName))
-		c.Check(plug.Name, Equals, plugName)
-		c.Check(plug.Hooks, HasLen, 1)
-		hook = plug.Hooks[hookName]
-		c.Assert(hook, NotNil, Commentf("Expected plug to be associated with hook %q", hookName))
-		c.Check(hook.Name, Equals, hookName)
-
-		// Verify also that the hook plug made it into info.Plugs
-		c.Check(info.Plugs[plugName], DeepEquals, plug)
 	}
 }
